MOD: rename locks_ to db_locks_ in storage

+++ src/config.cc
@@ -101,7 +101,7 @@ bool Config::parseConfigFromString(std::string input, std::string *err) {
     std::vector<std::string> loglevels = {"info", "warning", "error", "fatal"};
     for (size_t i = 0; i < loglevels.size(); i++) {
       if (Util::ToLower(args[1]) == loglevels[i]) {
-        loglevel = i;
+        loglevel = static_cast<int>(i);
         break;
       }
     }
